+++ storage-resize-images/functions/src/util.ts
@@ -1,12 +1,3 @@
-import * as path from "path";
-
-export const extractFileNameWithoutExtension = (
-  filePath: string,
-  ext: string
-) => {
-  return path.basename(filePath, ext);
-};
-
 export const startsWithArray = (
   userInputPaths: string[],
   imagePath: string